refactor sun.js, replace with Sol class
